NERSystem/InverseTextNormalization/requirements.txt
actionlib==1.12.0
allennlp==1.0.0
allennlp-models==1.0.0
appdirs==1.4.4
appnope==0.1.0
apturl==0.5.2
asn1crypto==0.24.0
attrs==19.1.0
backcall==0.1.0
beautifulsoup4==4.6.0
blis==0.4.1
boto3==1.14.34
botocore==1.17.34
Brlapi==0.6.6
catalogue==1.0.0
catkin==0.7.23
certifi==2019.3.9
chardet==3.0.4
Click==7.0
command-not-found==0.3
conllu==3.0
cryptography==2.1.4
cupshelpers==1.0
cv-bridge==1.13.0
cymem==2.0.3
Cython==0.29.15
dataclasses==0.7
decorator==4.4.0
defer==1.0.6
diagnostic-updater==1.9.3
distlib==0.3.0
distro-info===0.18ubuntu0.18.04.1
docutils==0.15.2
dynamic-reconfigure==1.6.3
filelock==3.0.12
Flask==1.0.2
Flask-SQLAlchemy==2.4.0
future==0.18.2
gencpp==0.6.5
geneus==2.2.6
genlisp==0.4.16
genmsg==0.5.15
gennodejs==2.0.1
genpy==0.6.9
geographiclib==1.50
geopy==2.0.0
gunicorn==19.9.0
h5py==2.10.0
html5lib==0.999999999
httplib2==0.9.2
idna==2.8
importlib-metadata==1.6.1
importlib-resources==2.0.1
iniconfig==1.0.1
iotop==0.6
ipython==7.5.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.13.3
Jinja2==2.10.1
jmespath==0.10.0
joblib==0.16.0
jsonnet==0.16.0
jsonpickle==1.4.1
jsonschema==3.0.1
jupyter-core==4.4.0
kazam==1.5.3
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
lark-parser==0.9.0
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
louis==3.5.0
lxml==4.2.1
macaroonbakery==1.1.3
Mako==1.0.7
Markdown==2.6.9
MarkupSafe==1.1.1
message-filters==1.14.5
more-itertools==8.4.0
murmurhash==1.0.2
nbformat==4.4.0
netifaces==0.10.4
nltk==3.5
nose==1.3.7
numpy==1.18.1
oauth==1.0.1
olefile==0.45.1
openshot-qt==2.4.4
overrides==3.0.0
packaging==20.4
parso==0.4.0
pexpect==4.7.0
pickleshare==0.7.5
Pillow==5.1.0
pipenv==2020.6.2
plac==1.1.3
plotly==3.9.0
pluggy==0.13.1
preshed==3.0.2
prompt-toolkit==2.0.9
protobuf==3.12.4
ptyprocess==0.6.0
py==1.9.0
py-rouge==1.1
pycairo==1.16.2
pycparser==2.19
pycrypto==2.6.1
pycups==1.9.73
Pygments==2.4.0
pygobject==3.26.1
pymacaroons==0.13.0
PyNaCl==1.1.2
pyparsing==2.4.7
pyRFC3339==1.0
pyrsistent==0.15.2
pytest==6.0.1
python-apt==1.6.5+ubuntu0.2
python-dateutil==2.6.1
python-debian==0.1.32
pyttsx3==2.81
pytz==2018.3
pyxdg==0.25
PyYAML==3.12
pyzmq==16.0.2
regex==2020.7.14
remarkable==1.87
reportlab==3.4.0
requests==2.21.0
requests-unixsocket==0.1.5
retrying==1.3.3
rosbag==1.14.5
rosclean==1.14.8
rosgraph==1.14.5
roslaunch==1.14.5
roslib==1.14.8
roslz4==1.14.5
rosmaster==1.14.5
rosmsg==1.14.5
rosnode==1.14.5
rosparam==1.14.5
rospy==1.14.5
rosservice==1.14.5
rostest==1.14.5
rostopic==1.14.5
rosunit==1.14.8
roswtf==1.14.5
s3transfer==0.3.3
sacremoses==0.0.43
scikit-learn==0.23.1
scipy==1.5.2
scour==0.36
SecretStorage==2.3.1
sensor-msgs==1.12.7
sentencepiece==0.1.91
simplejson==3.13.2
six==1.11.0
smclib==1.8.5
spacy==2.2.4
SQLAlchemy==1.3.3
srsly==1.0.2
system-service==0.3
systemd-python==234
tensorboardX==2.1
tf==1.12.1
tf2-py==0.6.5
tf2-ros==0.6.5
thinc==7.4.0
threadpoolctl==2.1.0
tokenizers==0.7.0
toml==0.10.1
topic-tools==1.14.5
torch==1.5.1
tqdm==4.48.2
traitlets==4.3.2
transformers==2.11.0
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.24.1
usb-creator==0.3.3
virtualenv==20.0.25
virtualenv-clone==0.5.4
vtk==8.1.2
wadllib==1.3.2
wasabi==0.7.1
wcwidth==0.1.7
webencodings==0.5
Werkzeug==0.15.2
word2number==1.1
xkit==0.0.0
zipp==3.1.0
zope.interface==4.3.2
